backend/pyproject.toml
[project]
name = "fastapitemplate"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.14"
readme = "README.md"
dependencies = [
    "fastapi",
    "dishka",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "greenlet",
    "pydantic-settings",
    "orjson",
    "hypercorn",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
extend-select = ["I"]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
]

ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.14"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
